Remove unused code and server files
