+++ src/server/sim.py
@@ -1,7 +1,6 @@
 import socket
 import threading
 import time
-import pickle
 import socket
 
 from src.core.util import date_range
@@ -10,10 +9,7 @@ from src.core.meter import mk_meters_runner
 def mksocket() -> socket.socket:
     return socket.socket(socket.AF_INET, socket.SOCK_STREAM)
 
-
-def make_simulation_server(n, server_address, ui_address):
-    ui_conn = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    ui_conn.connect(ui_address)
+def make_simulation_server_state(n, server_address, append_state):
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
         s.bind(server_address)
         s.listen(n)
@@ -30,7 +26,7 @@ def make_simulation_server(n, server_address, ui_address):
     def simulate(start_date, end_date, datetime_delta, refresh_rate, moment):
         for t in date_range(start_date, end_date, datetime_delta):
             state = moment(t, conns)
-            ui_conn.sendall(pickle.dumps(state))
+            append_state(state)
             time.sleep(refresh_rate)
 
     return simulate
